Nika: Basak's features

See merge request gagneurlab/gfeat!16

+++ gfeat/score_UTRs.py
@@ -7,9 +7,10 @@ from pyensembl import EnsemblRelease
 from cyvcf2 import VCF
 import pandas as pd
 import numpy as np
+import os
 
 
-def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, gtf=None, fasta=None):
+def score_utrs(vcf, save_to, patient_id, ensembl_version=None, gtf=None, fasta=None):
     """
     1 possible model built using gfeat
     Looks for creation and deletion of AUGs which are located upstream from the corresponding canonical start.
@@ -17,10 +18,12 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
     AUGType (in-frame_no_uORF, in-frame_uORF, not_in-frame_no_uORF, not_in-frame_uORF),
     alternative value of the nucleobase, chromosome, reference value of the nucleobase, sampleID
     :param vcf: string, path to the vcf.gz
-    :param save_extra_to_csv: bool, whether to save intermediate information (all AUGs, transcripts with too many
-                              mutations and some very basic statistics) to a csv file or not
-    :param path: string, path to the folder where to save the pd.DataFrame output table
+
+           currently work only if save_to is the name of the folder
+    :param save_to: string, path to the folder or the file name where to save the pd.DataFrame output table
     :param ensembl_version:, int, Ensembl version
+           MIN_ENSEMBL_RELEASE = 54
+           MAX_ENSEMBL_RELEASE = 85
     :param gtf: string, path to the gtf file, as an alternative to the Ensembl version, user gtf and Fasta files
                 can be provided
     :param fasta: string, path to the Fasta file, string, as an alternative to the Ensembl version, user gtf
@@ -49,27 +52,12 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
     for field in vcf_fields:
         vcf_file.add_info_to_header({"ID": field, "Number": 1, "Type": "String", "Description": "dummy"})
 
-    if save_extra_to_csv:
-        dictionary = {'Gene': [], 'Transcript': [], 'not_in-frame_no_uORF': [], 'not_in-frame_uORF': [],
-                      'in-frame_no_uORF': [], 'in-frame_uORF': []}
-        error_dictionary = {'More_than_14_mutations': []}
-        sum_dictionary = {'Contig': [], 'Total_transcript_num': [], 'Mutated_transcript_num': [],
-                          'not_in-frame_no_uORF_num': [],
-                          'not_in-frame_uORF': [], 'in-frame_no_uORF_num': [], 'in-frame_uORF_num': []}
-        mutated_dictionary = {'Gene': [], 'Transcript': [], 'Type': [], 'Position': [], 'ref': [],
-                              'alt': [], 'chr': [], 'sample': []}
-
-        df0 = pd.DataFrame(data=dictionary)
-        df1 = pd.DataFrame(data=error_dictionary)
-        df2 = pd.DataFrame(data=sum_dictionary)
-
-        df0.to_csv(path + "/" + patient_id + "_AUGs.csv")
-        df1.to_csv(path + "/" + patient_id + "_too_many_mutated.csv")
-        df2.to_csv(path + "/" + patient_id + "_summary.csv")
+    mutated_dictionary = {'Gene': [], 'Transcript': [], 'Type': [], 'Position': [], 'ref': [],
+                              'alt': [], 'chr': [], 'Sample': [], 'Creation': []}
 
     for contig in contigs:
 
-        ds = FivePrimeUTRSeq(data, False, contig, '+')  # Todo: check
+        ds = FivePrimeUTRSeq(data, False, contig, '+')
 
         dictionary = {'Gene': [], 'Transcript': [], 'not_in-frame_no_uORF': [], 'not_in-frame_uORF': [],
                       'in-frame_no_uORF': [], 'in-frame_uORF': []}
@@ -179,7 +167,11 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
                                         mutated_dictionary['ref'].append(info['ref'][i])
                                         mutated_dictionary['alt'].append(info['alt'][i])
                                         mutated_dictionary['chr'].append(contig)
-                                        mutated_dictionary['sample'].append(patient_id)
+                                        mutated_dictionary['Sample'].append(patient_id)
+                                        if pos in mut_0_0:
+                                            mutated_dictionary['Creation'].append(1)
+                                        else:
+                                            mutated_dictionary['Creation'].append(0)
                     mut_0_1 = dictionary['not_in-frame_uORF'][len(dictionary['not_in-frame_uORF']) - 1]
                     if not np.array_equal(mut_0_1, cur_0_1):
                         for pos in np.setdiff1d(np.union1d(mut_0_1, cur_0_1), np.intersect1d(mut_0_1, cur_0_1)):
@@ -196,7 +188,11 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
                                         mutated_dictionary['ref'].append(info['ref'][i])
                                         mutated_dictionary['alt'].append(info['alt'][i])
                                         mutated_dictionary['chr'].append(contig)
-                                        mutated_dictionary['sample'].append(patient_id)
+                                        mutated_dictionary['Sample'].append(patient_id)
+                                        if pos in mut_0_1:
+                                            mutated_dictionary['Creation'].append(1)
+                                        else:
+                                            mutated_dictionary['Creation'].append(0)
                     mut_1_0 = dictionary['in-frame_no_uORF'][len(dictionary['in-frame_no_uORF']) - 1]
                     if not np.array_equal(mut_1_0, cur_1_0):
                         for pos in np.setdiff1d(np.union1d(mut_1_0, cur_1_0), np.intersect1d(mut_1_0, cur_1_0)):
@@ -212,7 +208,11 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
                                         mutated_dictionary['ref'].append(info['ref'][i])
                                         mutated_dictionary['alt'].append(info['alt'][i])
                                         mutated_dictionary['chr'].append(contig)
-                                        mutated_dictionary['sample'].append(patient_id)
+                                        mutated_dictionary['Sample'].append(patient_id)
+                                        if pos in mut_1_0:
+                                            mutated_dictionary['Creation'].append(1)
+                                        else:
+                                            mutated_dictionary['Creation'].append(0)
                     mut_1_1 = dictionary['in-frame_uORF'][len(dictionary['in-frame_uORF']) - 1]
                     if not np.array_equal(mut_1_1, cur_1_1):
                         for pos in np.setdiff1d(np.union1d(mut_1_1, cur_1_1), np.intersect1d(mut_1_1, cur_1_1)):
@@ -228,21 +228,13 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
                                         mutated_dictionary['ref'].append(info['ref'][i])
                                         mutated_dictionary['alt'].append(info['alt'][i])
                                         mutated_dictionary['chr'].append(contig)
-                                        mutated_dictionary['sample'].append(patient_id)
+                                        mutated_dictionary['Sample'].append(patient_id)
+                                        if pos in mut_1_1:
+                                            mutated_dictionary['Creation'].append(1)
+                                        else:
+                                            mutated_dictionary['Creation'].append(0)
             else:
                 list_many_mut.append((ds[i])["transcripts"][0])
-
-        if save_extra_to_csv:
-            df0 = pd.DataFrame(data=dictionary)
-            error_list = list_many_mut
-            error_dictionary = {"More_than_14_mutations": error_list}
-            df1 = pd.DataFrame(data=error_dictionary)
-
-            with open(path + "/" + patient_id + "_AUGs.csv", 'a') as f:
-                df0.to_csv(f, header=False)
-
-            with open(path + "/" + patient_id + "_too_many_mutated.csv", 'a') as f:
-                df1.to_csv(f, header=False)
 
         ds = FivePrimeUTRSeq(data, False, contig, '-')
 
@@ -353,7 +345,11 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
                                         mutated_dictionary['ref'].append(info['ref'][i])
                                         mutated_dictionary['alt'].append(info['alt'][i])
                                         mutated_dictionary['chr'].append(contig)
-                                        mutated_dictionary['sample'].append(patient_id)
+                                        mutated_dictionary['Sample'].append(patient_id)
+                                        if pos in mut_0_0:
+                                            mutated_dictionary['Creation'].append(1)
+                                        else:
+                                            mutated_dictionary['Creation'].append(0)
                     mut_0_1 = dictionary['not_in-frame_uORF'][len(dictionary['not_in-frame_uORF']) - 1]
                     if not np.array_equal(mut_0_1, cur_0_1):
                         for pos in np.setdiff1d(np.union1d(mut_0_1, cur_0_1), np.intersect1d(mut_0_1, cur_0_1)):
@@ -370,7 +366,11 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
                                         mutated_dictionary['ref'].append(info['ref'][i])
                                         mutated_dictionary['alt'].append(info['alt'][i])
                                         mutated_dictionary['chr'].append(contig)
-                                        mutated_dictionary['sample'].append(patient_id)
+                                        mutated_dictionary['Sample'].append(patient_id)
+                                        if pos in mut_0_1:
+                                            mutated_dictionary['Creation'].append(1)
+                                        else:
+                                            mutated_dictionary['Creation'].append(0)
                     mut_1_0 = dictionary['in-frame_no_uORF'][len(dictionary['in-frame_no_uORF']) - 1]
                     if not np.array_equal(mut_1_0, cur_1_0):
                         for pos in np.setdiff1d(np.union1d(mut_1_0, cur_1_0), np.intersect1d(mut_1_0, cur_1_0)):
@@ -386,7 +386,11 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
                                         mutated_dictionary['ref'].append(info['ref'][i])
                                         mutated_dictionary['alt'].append(info['alt'][i])
                                         mutated_dictionary['chr'].append(contig)
-                                        mutated_dictionary['sample'].append(patient_id)
+                                        mutated_dictionary['Sample'].append(patient_id)
+                                        if pos in mut_1_0:
+                                            mutated_dictionary['Creation'].append(1)
+                                        else:
+                                            mutated_dictionary['Creation'].append(0)
                     mut_1_1 = dictionary['in-frame_uORF'][len(dictionary['in-frame_uORF']) - 1]
                     if not np.array_equal(mut_1_1, cur_1_1):
                         for pos in np.setdiff1d(np.union1d(mut_1_1, cur_1_1), np.intersect1d(mut_1_1, cur_1_1)):
@@ -402,29 +406,25 @@ def score_utrs(vcf, save_extra_to_csv, path, patient_id, ensembl_version=None, g
                                         mutated_dictionary['ref'].append(info['ref'][i])
                                         mutated_dictionary['alt'].append(info['alt'][i])
                                         mutated_dictionary['chr'].append(contig)
-                                        mutated_dictionary['sample'].append(patient_id)
+                                        mutated_dictionary['Sample'].append(patient_id)
+                                        if pos in mut_1_1:
+                                            mutated_dictionary['Creation'].append(1)
+                                        else:
+                                            mutated_dictionary['Creation'].append(0)
             else:
                 list_many_mut.append((ds[i])["transcripts"][0])
 
-        if save_extra_to_csv:
-            df0 = pd.DataFrame(data=dictionary)
-            error_list = list_many_mut
-            error_dictionary = {"More_than_14_mutations": error_list}
-            df2 = pd.DataFrame(data=sum_dictionary)
-            df1 = pd.DataFrame(data=error_dictionary)
-
-            with open(path + "/" + patient_id + "_AUGs.csv", 'a') as f:
-                df0.to_csv(f, header=False)
-
-            with open(path + "/" + patient_id + "_too_many_mutated.csv", 'a') as f:
-                df1.to_csv(f, header=False)
-
-            with open(path + "/" + patient_id + "_summary.csv", 'a') as f:
-                df2.to_csv(f, header=False)
-
     df3 = pd.DataFrame(data=mutated_dictionary)
     df3 = df3.iloc[df3.astype(str).drop_duplicates().index]
-    file = path + "/mutated.csv"
-    with open(file, 'a') as f:
-        df3.to_csv(f, header=False)
+
+    if save_to.endswith('.csv'):
+        file = save_to
+    else:
+        file = save_to + "/" + patient_id + "_mutated.csv"
+
+    if os.path.isfile(file):
+        with open(file, 'a') as f:
+            df3.to_csv(f, header=False)
+    else:
+        df3.to_csv(file)
     pass
